[Booking] List your items section - bug fixing
The height of the box for the whole list of items section adjust  close to the footer.
add button initially use the disabled state (#DFDFDF) if nothing is being entered to the item name yet.
Proceed button adjust place far from the add button and almost at the bottom.
The supporting text under on the left changed to body size